Restructure and generalize views and tests
Generic views and test cases have been moved to the `base` app.
