+++ liquidator/settings.py
@@ -40,6 +40,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'corsheaders',
+    'base',
     'company',
     'custom_auth',
     'transaction',
